requirements.txt
Flask
Flask_SQLAlchemy
Authlib
stripe
python-dotenv
gunicorn
